+++ print_binary.c
@@ -1,12 +1,12 @@
 #include "main.h"
 
 /**
- * print_binary - prints an unsigned int as binary
+ * print_binary - prints an unsigned long int as binary
  * @n: number to print
  *
  * Return: number of characters printed
  */
-int print_binary(unsigned int n)
+int print_binary(unsigned long int n)
 {
 	int count = 0;
 
@@ -14,5 +14,6 @@ int print_binary(unsigned int n)
 		count += print_binary(n / 2);
 
 	count += _putchar((n % 2) + '0');
+
 	return (count);
 }
